refactor(playground-ui): consolidate SyntaxHighlighter components
- Create CodeEditor component in ds/components for editable JSON/code display
  - Merges features from previous implementations
  - Uses proper design tokens (bg-surface4) instead of hardcoded colors
  - Includes CopyButton, className support, and useCodemirrorTheme hook

- Create SyntaxHighlighter component in ds/components for read-only markdown
  - Prism-based for @assistant-ui/react-markdown integration
  - Used for rendering code blocks in AI assistant responses

- Remove duplicate implementations:
  - src/components/syntax-highlighter.tsx
  - src/components/ui/syntax-highlighter.tsx
  - src/components/assistant-ui/syntax-highlighter.tsx

- Update all consuming files (15 files) to use new component locations
