+++ packages/playground-ui/src/components/assistant-ui/messages/markdown-text.tsx
@@ -10,18 +10,9 @@ import '@assistant-ui/react-markdown/styles/dot.css';
 import { CheckIcon, CopyIcon } from 'lucide-react';
 import { FC, ImgHTMLAttributes, memo, useEffect, useState } from 'react';
 import remarkGfm from 'remark-gfm';
-import { makePrismAsyncLightSyntaxHighlighter } from '@assistant-ui/react-syntax-highlighter';
-import { coldarkDark } from 'react-syntax-highlighter/dist/cjs/styles/prism';
 import { TooltipIconButton } from '@/components/assistant-ui/tooltip-icon-button';
 import { cn } from '@/lib/utils';
-
-const SyntaxHighlighter = makePrismAsyncLightSyntaxHighlighter({
-  style: coldarkDark,
-  customStyle: {
-    margin: 0,
-    backgroundColor: 'black',
-  },
-});
+import { SyntaxHighlighter } from '@/ds/components/SyntaxHighlighter';
 
 const MarkdownTextImpl = () => {
   return <MarkdownTextPrimitive remarkPlugins={[remarkGfm]} className="aui-md" components={defaultComponents} />;
