+++ packages/playground-ui/src/components/assistant-ui/tools/badges/tool-badge.tsx
@@ -2,7 +2,7 @@ import { ToolsIcon } from '@/ds/icons';
 import { Button } from '@/ds/components/Button';
 import { Icon } from '@/ds/icons';
 import { Check, X } from 'lucide-react';
-import { SyntaxHighlighter } from '../../../ui/syntax-highlighter';
+import { CodeEditor } from '@/ds/components/CodeEditor';
 import { BadgeWrapper } from './badge-wrapper';
 import { NetworkChoiceMetadataDialogTrigger } from './network-choice-metadata-dialog';
 import { MastraUIMessage } from '@mastra/react';
@@ -31,7 +31,7 @@ export const ToolBadge = ({
 
   try {
     const { __mastraMetadata: _, ...formattedArgs } = typeof args === 'object' ? args : JSON.parse(args);
-    argSlot = <SyntaxHighlighter data={formattedArgs} data-testid="tool-args" />;
+    argSlot = <CodeEditor data={formattedArgs} data-testid="tool-args" />;
   } catch {
     argSlot = <pre className="whitespace-pre bg-surface4 p-4 rounded-md overflow-x-auto">{args as string}</pre>;
   }
@@ -40,14 +40,14 @@ export const ToolBadge = ({
     typeof result === 'string' ? (
       <pre className="whitespace-pre bg-surface4 p-4 rounded-md overflow-x-auto">{result}</pre>
     ) : (
-      <SyntaxHighlighter data={result} data-testid="tool-result" />
+      <CodeEditor data={result} data-testid="tool-result" />
     );
 
   let suspendPayloadSlot =
     typeof suspendPayload === 'string' ? (
       <pre className="whitespace-pre bg-surface4 p-4 rounded-md overflow-x-auto">{suspendPayload}</pre>
     ) : (
-      <SyntaxHighlighter data={suspendPayload} data-testid="tool-suspend-payload" />
+      <CodeEditor data={suspendPayload} data-testid="tool-suspend-payload" />
     );
 
   const selectionReason = metadata?.mode === 'network' ? metadata.selectionReason : undefined;
@@ -95,7 +95,7 @@ export const ToolBadge = ({
             <p className="font-medium pb-2">Tool output</p>
 
             <div className="h-40 overflow-y-auto">
-              <SyntaxHighlighter data={toolOutput} data-testid="tool-output" />
+              <CodeEditor data={toolOutput} data-testid="tool-output" />
             </div>
           </div>
         )}
