+++ packages/playground-ui/src/domains/observability/components/tracing-run-options.tsx
@@ -1,6 +1,6 @@
 import CodeMirror from '@uiw/react-codemirror';
 import { jsonLanguage } from '@codemirror/lang-json';
-import { useCodemirrorTheme } from '@/components/syntax-highlighter';
+import { useCodemirrorTheme } from '@/ds/components/CodeEditor';
 import { useTracingSettings } from '@/domains/observability/context/tracing-settings-context';
 import { Txt } from '@/ds/components/Txt';
 import { WorkflowRunOptions } from '@/domains/workflows/workflow/workflow-run-options';
