+++ packages/playground-ui/src/ds/components/SyntaxHighlighter/SyntaxHighlighter.tsx
