+++ packages/playground-ui/src/ds/components/SyntaxHighlighter/index.ts
@@ -1,0 +1,1 @@
+export * from './SyntaxHighlighter';
